+++ CHANGELOG.md
@@ -19,3 +19,5 @@
    - use hex keys to storage resources (long_id not good for umsgpack)
 - [enhancement] create multi-peer working examples (examples/multi_peer_simulator.py)
 - [enhancement] add tcp listener on server
+   - remove Datagram abstraction
+   - remove RPCMessageQueue abstraction